Support Babelify (and other transforms) running 1st
Falafel uses acorn, which fails to handle ES6 import/export.
By running the cmify transform after babelify, Falafel (acorn) handle the
transformed code just fine.

This will also allow other transforms to take place before cmify replaces the
loads (such as rewriting paths, etc, if necessary).
