+++ plugin.js
@@ -2,8 +2,8 @@
 
 const cmify = require('./index')
 const through = require('through2')
-const falafel = require('falafel')
 const str = require('string-to-stream')
+const createCssModuleSource = require('./create-css-module-source')
 
 function createCmStream () {
   // fake module
@@ -23,76 +23,6 @@ cmify.getAllCss = function () { return ${JSON.stringify(cmify.getAllCss())} }
 module.exports = cmify`
 }
 
-function createCssModuleSource (filename) {
-  const tokens = cmify.load(filename)
-
-  // make sure all dependencies are added to browserify's tree
-  const output = tokens._deps.map(function (f) {
-    return 'require("' + f + '")'
-  })
-
-  // export the css module's tokens
-  output.push(`/** last updated: ${Date.now()} **/`)
-  output.push(`module.exports=${JSON.stringify(tokens)}`)
-  return output.join('\n')
-}
-
-function cmifyTransform (filename) {
-  const bufs = []
-  let cmifyName = null
-
-  const stream = through(write, end)
-  return stream
-
-  // ----
-
-  function write (buf, enc, next) {
-    bufs.push(buf)
-    next()
-  }
-
-  function end (done) {
-    const src = Buffer.concat(bufs).toString('utf8')
-
-    // handle css files
-    if (/\.css$/.test(filename)) {
-      try {
-        this.push(createCssModuleSource(filename))
-      }
-      catch (err) {
-        this.emit("error", err);
-      }
-    } else {
-      const ast = falafel(src, { ecmaVersion: 6 }, walk)
-      this.push(ast.toString())
-    }
-
-    this.push(null)
-    done()
-  }
-
-  function walk (node) {
-    // find `require('cmify')` and record the name it's bound to
-    if (node.type === 'CallExpression' &&
-        node.callee && node.callee.name === 'require' &&
-        node.arguments.length === 1 &&
-        node.arguments[0].value === 'cmify') {
-      cmifyName = node.parent.id.name
-      return
-    }
-
-    // find places where `cmify.load(...)` is called
-    if (node.type === 'CallExpression' &&
-        node.callee && node.callee.type === 'MemberExpression' &&
-        node.callee.object.name === cmifyName &&
-        node.callee.property.name === 'load'
-       ) {
-      // rewrite as `require`, so it gets included in the dependency tree
-      node.update(`require(${node.arguments[0].raw})`)
-    }
-  }
-}
-
 function cmifyPlugin (b, opts) {
   opts = opts || {}
 
@@ -104,8 +34,6 @@ function cmifyPlugin (b, opts) {
   // register a fake cmify module for the browser
   const cmStream = createCmStream()
   b.require(cmStream, { expose: cmStream.id })
-
-  b.transform(cmifyTransform)
 
   b.on('reset', reset)
   reset()
